Some changes to IDs/function names
Changed some names to the quizQuestions function to 'startQuiz' because I want that function to start it and then have another function or inside the same function that let's it go to the next question if a button has been pressed. However, now nothing happens when I click on the start quiz button, but I think it's because one variable is the same name as the function now. Gonna change the name of the variable "startQuiz" to something else and try again.