Changes
Buttons answers changes with innerText which was nice to see, using index++ in the 'startQuiz' function. My other function 'quizQuestions' is not working as intended yet and if I use it at the end of the 'startQuiz' function, the score in the end stays at "1" and it does not change no matter which button was clicked so I have to correct it or try to make it work in the same function instead. Continuing.