Just thinking/reading. No real changes.
Spent more time reading through the lesson notes. My buttons are still shown as undefined with the 'userClick' parameter/variable so I need to assign it somehow. The final button I click doesn't show up as undefined but the ones before that do. Confused but trying to fix it and hopefully tomorrow I can ask Collin for tips/hints.