Fixed bug(s)! Big improvement!
I spent SO many hours today trying to fix the previous problem in so many ways but in the end I used event in my function parameter and used the property 'event.target' with .innerText to track the score/answers correctly! I was stressed and this is now a huge relief! Super happy with it. Working on the restart button function next that I want to show up after the last question and answer button has been clicked. Then I will go on and change the current questions/answers/correct answers to 'real' ones instead of the ones I have now for testing.