all questions/buttons generate random numbers
Made all of the current buttons and questions generate random numbers but unfortunately I did it in a inconvenient way but at least it works. I still don't know how I am gonna change the click sequence, especially the way it's looking now but I am still happy with it!